# Add an agent tool for running a query