--- a/src/constants/constants.ts
+++ b/src/constants/constants.ts
@@ -182,6 +182,7 @@ export const copilotListTablesToolName = "mssql_list_tables";
 export const copilotListSchemasToolName = "mssql_list_schemas";
 export const copilotListViewsToolName = "mssql_list_views";
 export const copilotListFunctionsToolName = "mssql_list_functions";
+export const copilotRunQueryToolName = "mssql_run_query";
 export const copilotChangeDatabaseToolName = "mssql_change_database";
 export const copilotShowSchemaToolName = "mssql_show_schema";
 export const copilotGetConnectionDetailsToolName = "mssql_get_connection_details";