--- a/src/copilot/tools/listTablesTool.ts
+++ b/src/copilot/tools/listTablesTool.ts
@@ -11,7 +11,7 @@ import { MssqlChatAgent as loc } from "../../constants/locConstants";
 import { getErrorMessage } from "../../utils/utils";
 import SqlToolsServiceClient from "../../languageservice/serviceclient";
 import { RequestType } from "vscode-languageclient";
-import { SimpleExecuteResult } from "../interfaces";
+import { SimpleExecuteResult } from "vscode-mssql";
 import { listTablesQuery } from "../queries";
 
 export interface ListTablesToolParams {