… src/copilot/interfaces.ts